# Reduce argument overhead