--- a/lib/graphql/schema/field.rb
+++ b/lib/graphql/schema/field.rb
@@ -650,27 +650,29 @@ module GraphQL
             arg_values = args
             using_arg_values = false
           end
-          # Faster than `.any?`
-          arguments(context).each_value do |arg|
-            arg_key = arg.keyword
-            if arg_values.key?(arg_key)
-              arg_value = arg_values[arg_key]
-              if using_arg_values
-                if arg_value.default_used?
-                  # pass -- no auth required for default used
-                  next
-                else
-                  application_arg_value = arg_value.value
-                  if application_arg_value.is_a?(GraphQL::Execution::Interpreter::Arguments)
-                    application_arg_value.keyword_arguments
+          if args.size > 0
+            args = context.warden.arguments(self)
+            args.each do |arg|
+              arg_key = arg.keyword
+              if arg_values.key?(arg_key)
+                arg_value = arg_values[arg_key]
+                if using_arg_values
+                  if arg_value.default_used?
+                    # pass -- no auth required for default used
+                    next
+                  else
+                    application_arg_value = arg_value.value
+                    if application_arg_value.is_a?(GraphQL::Execution::Interpreter::Arguments)
+                      application_arg_value.keyword_arguments
+                    end
                   end
+                else
+                  application_arg_value = arg_value
                 end
-              else
-                application_arg_value = arg_value
-              end
 
-              if !arg.authorized?(object, application_arg_value, context)
-                return false
+                if !arg.authorized?(object, application_arg_value, context)
+                  return false
+                end
               end
             end
           end