Looking at a profile, I found that `arguments(ctx)` hashes are merged over and over again at runtime, at least in these two places. We should cache these per-query, too.

```diff 
-                           3.335  (± 0.0%) i/s -     17.000  in   5.101803s
+                           3.532  (± 0.0%) i/s -     18.000  in   5.097713s


-      55774   (8.8%)       10019   (1.6%)     GraphQL::Schema::Member::HasArguments#arguments
-      93425  (14.7%)        5604   (0.9%)     GraphQL::Schema::Member::HasArguments#coerce_arguments
-      18269   (2.9%)        4890   (0.8%)     GraphQL::Schema::Warden#visible_argument?
-      26798   (4.2%)        3665   (0.6%)     GraphQL::Schema::Field#authorized?

+      58446  (10.2%)        5725   (1.0%)     GraphQL::Schema::Member::HasArguments#coerce_arguments
+      22136   (3.9%)        4406   (0.8%)     GraphQL::Schema::Member::HasArguments#arguments
+       7366   (1.3%)        2176   (0.4%)     GraphQL::Schema::Warden#visible_argument?
+       4344   (0.8%)        2673   (0.5%)     GraphQL::Schema::Field#authorized?


- Total allocated: 16748432 bytes (105490 objects)
+ Total allocated: 15404096 bytes (97488 objects)
```

TODO: 

- [x] Post benchmark improvements
- [ ] Port to 2.0.x

I also investigated caching coerced arguments in 1484b84d6, but I didn't see any further improvement in the introspection benchmark.