@rmosolgo 2023-02-14
Looking into a big introspection benchmark, it looks like this reduced time in `HasArguments#arguments` by about half (3% of total query time):


  | Total | Self | Method
-- | -- | -- | --
Before | 1.17s (5.6%) | 259.05ms (1.2%) | GraphQL::Schema::Member::HasArguments#arguments
After | 561.36ms (2.6%) | 201.18ms (0.93%) | GraphQL::Schema::Member::HasArguments#arguments